+++ URLShortener.Domain/Entities/Finance/PurchaseModel.cs
@@ -4,15 +4,15 @@ namespace URLShortener.Domain.Entities.Finance;
 
 public class PurchaseModel
 {
-    public int ID { get; set; }
+    public long ID { get; set; }
     public DateTime CreatedAt { get; set; }
     public long Amount { get; set; }
 
     public required int ServiceType { get; set; }
 
-    public int CustomURLID { get; set; }
+    public long CustomURLID { get; set; }
     public required URLModel URL { get; set; }
 
-    public int FinanceID { get; set; }
+    public long FinanceID { get; set; }
     public required FinancialRecordModel Finance { get; set; }
 }
