+++ URLShortener.Domain/Entities/URL/URLModel.cs
@@ -8,7 +8,7 @@ namespace URLShortener.Domain.Entities.URL;
 
 public class URLModel
 {
-    public int ID { get; set; }
+    public long ID { get; set; }
     public string? Description { get; set; }
     public required string ShortCode { get; set; }
     public required string LongURL { get; set; }
@@ -17,17 +17,17 @@ public class URLModel
     public int ClickCount { get; set; }
     public bool IsMonetized { get; set; }
 
-    public int UserID { get; set; }
+    public long UserID { get; set; }
     public required UserModel User { get; set; }
 
     public ICollection<ClickInfoModel>? Clicks { get; set; } = [];
 
-    public int? CategoryID { get; set; }
+    public long? CategoryID { get; set; }
     public ICollection<URLCategoryModel>? Categories { get; set; }
 
-    public int URLAnalyticsID { get; set; }
+    public long URLAnalyticsID { get; set; }
     public URLAnalyticsModel? URLAnalytics { get; set; }
 
-    public int? PurchaseID { get; set; }
+    public long? PurchaseID { get; set; }
     public PurchaseModel? Purchase { get; set; }
 }
