+++ URLShortener.Domain/Entities/URLCategory/URLCategoryModel.cs
@@ -5,11 +5,11 @@ namespace URLShortener.Domain.Entities.URLCategory;
 
 public sealed class URLCategoryModel
 {
-    public int ID { get; set; }
+    public long ID { get; set; }
     public required string Title { get; set; }
     public string? Description { get; set; }
 
-    public int UserID { get; set; }
+    public long UserID { get; set; }
     public UserModel? User { get; set; }
     public ICollection<URLModel>? URLs { get; set; }
 }
